torneosite/requirements.txt
Django==1.7
Pillow==2.6.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap-form==3.1
django-bootstrap3-datetimepicker==2.2.3
django-braces==1.4.0
django-gmapi==1.0.1
django-messages==0.5.1
django-simple-captcha==0.4.2
django-simple-email-confirmation==0.11
django-taggit==0.12.2
django-toolbelt==0.0.1
geopy==1.1.3
gunicorn==19.1.1
psycopg2==2.5.4
pytz==2014.7
selenium==2.42.1
simplejson==3.6.3
six==1.7.3
static3==0.5.1
wsgiref==0.1.2
